feat(cat-voices): voices drawer - side sheet (#3045)
* feat: change shape of voices drawer

* feat: voices drawer

* chore: drawer

* chore: cleanup
